Tests/LibWeb/Text/expected/selection-drag-below-line-with-side-content.txt

selected previous text: false
selected rest of line: true
selected sidebar text: false
